break out EstablishView submap
this encapsulates the machine to establish a missing view, genesis or
otherwise
