+++ consortium/src/main/java/com/salesforce/apollo/consortium/fsm/CollaboratorFsm.java
@@ -6,8 +6,6 @@
  */
 package com.salesforce.apollo.consortium.fsm;
 
-import java.util.List;
-
 import com.chiralbehaviors.tron.Entry;
 import com.chiralbehaviors.tron.Exit;
 import com.salesfoce.apollo.consortium.proto.Block;
@@ -18,10 +16,8 @@ import com.salesfoce.apollo.consortium.proto.Sync;
 import com.salesfoce.apollo.consortium.proto.TotalOrdering;
 import com.salesfoce.apollo.consortium.proto.Transaction;
 import com.salesfoce.apollo.consortium.proto.Validate;
-import com.salesforce.apollo.consortium.Consortium.CollaboratorContext;
 import com.salesforce.apollo.consortium.Consortium.Timers;
 import com.salesforce.apollo.consortium.CurrentBlock;
-import com.salesforce.apollo.consortium.EnqueuedTransaction;
 import com.salesforce.apollo.membership.Member;
 
 /**
@@ -32,52 +28,6 @@ import com.salesforce.apollo.membership.Member;
  */
 public enum CollaboratorFsm implements Transitions {
 
-    BUILD_VIEW {
-        @Exit
-        public void cancel() {
-            context().cancel(Timers.AWAIT_VIEW_MEMBERS);
-        }
-
-        @Override
-        public Transitions deliverStop(Stop stop, Member from) {
-            context().deliverStop(stop, from);
-            return null;
-        }
-
-        @Override
-        public Transitions deliverStopData(StopData stopData, Member from) {
-            context().deliverStopData(stopData, from);
-            return null;
-        }
-
-        @Override
-        public Transitions deliverTransaction(Transaction txn, Member from) {
-            context().receive(txn);
-            return null;
-        }
-
-        @Override
-        public Transitions deliverTransactions(ReplicateTransactions txns, Member from) {
-            context().receive(txns, from);
-            return null;
-        }
-
-        @Override
-        public Transitions formView() {
-            return VIEW_ESTABLISHED;
-        }
-
-        @Entry
-        public void petitionToJoin() {
-            context().joinView();
-        }
-
-        @Override
-        public Transitions receive(Transaction transacton, Member from) {
-            context().receive(transacton);
-            return null;
-        }
-    },
     CLIENT {
     },
     FOLLOWER
@@ -314,13 +264,8 @@ public enum CollaboratorFsm implements Transitions {
 
         @Override
         public Transitions join() {
-            fsm().push(BUILD_VIEW);
+            fsm().push(EstablishView.BUILD);
             return null;
-        }
-
-        @Override
-        public Transitions joinGenesis() {
-            return BUILD_VIEW;
         }
 
         @Override
@@ -332,69 +277,6 @@ public enum CollaboratorFsm implements Transitions {
         @Override
         public Transitions processGenesis(CurrentBlock next) {
             context().processGenesis(next);
-            return null;
-        }
-    },
-    VIEW_ESTABLISHED {
-
-        @Override
-        public Transitions becomeFollower() {
-            fsm().pop().becomeFollower();
-            return FOLLOWER;
-        }
-
-        @Override
-        public Transitions becomeLeader() {
-            fsm().pop().becomeLeader();
-            return LEADER;
-        }
-
-        @Override
-        public Transitions deliverStop(Stop stop, Member from) {
-            context().deliverStop(stop, from);
-            return null;
-        }
-
-        @Override
-        public Transitions deliverStopData(StopData stopData, Member from) {
-            context().deliverStopData(stopData, from);
-            return null;
-        }
-
-        @Override
-        public Transitions deliverSync(Sync syncData, Member from) {
-            context().deliverSync(syncData, from);
-            return null;
-        }
-
-        @Override
-        public Transitions deliverTransaction(Transaction txn, Member from) {
-            context().receive(txn);
-            return null;
-        }
-
-        @Override
-        public Transitions deliverTransactions(ReplicateTransactions txns, Member from) {
-            context().receive(txns, from);
-            return null;
-        }
-
-        @Override
-        public Transitions deliverValidate(Validate validation) {
-            context().validate(validation);
-            return null;
-        }
-
-        @Override
-        public Transitions receive(Transaction transaction, Member from) {
-            context().receive(transaction);
-            return null;
-        }
-
-        @Override
-        public Transitions startRegencyChange(List<EnqueuedTransaction> transactions) {
-            CollaboratorContext context = context();
-            fsm().push(ChangeRegency.INITIAL, () -> context.changeRegency(transactions));
             return null;
         }
     };
